fix: change user_id and tester_id fields to use unsigned integers in user_tester_assignments table migration

+++ database/migrations/2026_04_02_000117_create_user_tester_assignments_table.php
@@ -12,18 +12,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('user_tester_assignments', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('tester_id');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('tester_id');
 
             $table->primary(['user_id', 'tester_id']);
 
             $table->foreign('user_id')
-                  ->references('id')->on('users')
-                  ->cascadeOnDelete();
+                ->references('id')->on('users')
+                ->cascadeOnDelete();
 
             $table->foreign('tester_id')
-                  ->references('id')->on('testers')
-                  ->cascadeOnDelete();
+                ->references('id')->on('testers')
+                ->cascadeOnDelete();
         });
     }
